Fix blank editor being converted to word "null"
Internal function htmlconvert() used to return null when the editor was
empty.  Unfortunately this cause IE to display the word "null" when
writing this to InnerHTML.  It now returns the empty string when the
editor is empty as the simplest solution.  Need to watch out for side
effects this may have.